divided up the code into multiple files
now there is a file for nick's functions, a file for micah's fuctions,
and a file for shared functions
