+++ PROS/include/mutual.h
@@ -1,0 +1,4 @@
+void tip();
+void lift(int direction);
+void autoButton();
+void controlSwitch(int controller);
