+++ PROS/include/nick.h
@@ -1,0 +1,3 @@
+void nDrive();
+void nLift();
+void nClaw();
