+++ PROS/src/init.c
@@ -22,7 +22,6 @@
  */
 void initializeIO() {
   // configure the pin that the limit swith is plugged into as an input
-    pinMode(LIMIT_SWITCH, INPUT);
 }
 
 /*
