+++ PROS/src/opcontrol.c
@@ -1,177 +1,40 @@
-/** @file opcontrol.c
- * @brief File for operator control code
- *
- * This file should contain the user operatorControl() function and any functions related to it.
- *
- * Any copyright is dedicated to the Public Domain.
- * http://creativecommons.org/publicdomain/zero/1.0/
- *
- * PROS contains FreeRTOS (http://www.freertos.org) whose source code may be
- * obtained from http://sourceforge.net/projects/freertos/files/ or on request.
- */
-
 #include "main.h"
-
-/*
- * Runs the user operator control code. This function will be started in its own task with the
- * default priority and stack size whenever the robot is enabled via the Field Management System
- * or the VEX Competition Switch in the operator control mode. If the robot is disabled or
- * communications is lost, the operator control task will be stopped by the kernel. Re-enabling
- * the robot will restart the task, not resume it from where it left off.
- *
- * If no VEX Competition Switch or Field Management system is plugged in, the VEX Cortex will
- * run the operator control task. Be warned that this will also occur if the VEX Cortex is
- * tethered directly to a computer via the USB A to A cable without any VEX Joystick attached.
- *
- * Code running in this task can take almost any action, as the VEX Joystick is available and
- * the scheduler is operational. However, proper use of delay() or taskDelayUntil() is highly
- * recommended to give other tasks (including system tasks such as updating LCDs) time to run.
- *
- * This task should never exit; it should end with some kind of infinite loop, even if empty.
- */
-
-
-//function for the lift arm
-void lift(int direction){
-	if (direction < 0 && analogRead(1) > 1850){
-		motorStop(4);
-		motorStop(5);
-		motorStop(6);
-		motorStop(7);
-	}else if(direction > -15
-		 && analogRead(1) < 50){
-		motorStop(4);
-		motorStop(5);
-		motorStop(6);
-		motorStop(7);
-	}else if(direction < 0 && analogRead(1) > 1200){
-		direction = direction + 20;
-		motorSet(4, direction); // set arm left 1
-		motorSet(5, direction); // set arm left 2
-		motorSet(6, direction); // set arm left 3
-		motorSet(7, direction); // set arm left 4
-	}else{
-		motorSet(4, direction); // set arm left 1
-		motorSet(5, direction); // set arm left 2
-		motorSet(6, direction); // set arm left 3
-		motorSet(7, direction); // set arm left 4
-	}
-}
-
-//tip-bar code
-void tipSet(int direction){
-	motorSet(1, direction); // main tip motor
-}
+#include "mutual.h"
+#include "micah.h"
+#include "nick.h"
 
 //main operator control statement
 void operatorControl() {
-	int power;
-  int turn;
-	int controller = 0;
-	//int clawDirection; // current movement of the claw
-	// 0 = stop, 1 = close, 2 = open
-    while (1) {
+
+	int controller = 0; // 0 = nick's controls, 1 = micah's
+
+	//main loop
+		while (1) {
+
+
+			//n fuctions are in nick.c
+			//m functions are in micah.c
+			
+			//check who's controls to use
 			if (controller == 1){
-				//drive code
-        power = joystickGetAnalog(1, 1); // vertical axis on left joystick
-        turn  = joystickGetAnalog(1, 2); // horizontal axis on left joystick
-				if(power > 15 || power < -15 || turn > 15 || turn < -15){
-	        motorSet(2, power + turn); // set left wheels
-					motorSet(9, power - turn); // set left wheels
-	        motorSet(3, power - turn); // set right
-					motorSet(10, power + turn); // set right
-				}else{
-					motorStop(2);
-					motorStop(3);
-					motorStop(9);
-					motorStop(10);
-				}
-
-
-				//lift code
-				lift(joystickGetAnalog(1, 3) - 8);
-
-
-				//tip-bar code
-				if(joystickGetDigital(1, 7, JOY_UP)) {
-		      tipSet(127); // pressing up, so tip-bar should open
-		    }else if(joystickGetDigital(1, 7, JOY_DOWN)) {
-		      tipSet(-127); // pressing down, so tip-bar should close
-		    }else {
-		      tipSet(0); // no buttons are pressed, stop the tip-bar
-		    }
-
-				/*not using potentiometer right now
-				//claw
-				clawDirection = clawGetDirection(); // get the input of the controller and set direction
-			  clawMove(clawDirection); // move claw based on direction
-				*/
-				//alternate claw code
-				if(joystickGetDigital(1, 6, JOY_UP)) {
-		      motorSet(8, 127);
-		    }else if(joystickGetDigital(1, 6, JOY_DOWN)) {
-		      motorSet(8, -127);
-		    }else{
-		      motorSet(8, 0);
-		    }
-
-
-				//button on the far right to trigger autonomous when no competition switch is connected
-				if(joystickGetDigital(1, 8, JOY_RIGHT) && isOnline() == false) {
-		      autonomous(); // this calls the autonomous code
-		    }
-
-				//control switch
-				if(joystickGetDigital(1, 8, JOY_UP)){
-					controller = 0;
-					delay(500);
-				}
-
+				// micah specific controls
+				mDrive();
+				lift(joystickGetAnalog(1, 3) - 8); // micah lift control
+				mClaw(); // claw control
 			}else{
-				//nick's control
-
-				//driving
-				int right = -joystickGetAnalog(1, 2);
-				int left = joystickGetAnalog(1, 3);
-				//set motors
-				motorSet(2, left);
-				motorSet(10, left);
-				motorSet(3, right);
-				motorSet(9, right);
-
-				//arm
-				if(joystickGetDigital(1, 6, JOY_UP)){
-					lift(-127);
-				}else if(joystickGetDigital(1, 6, JOY_DOWN)){
-					lift(127);
-				}else{
-					lift(0);
-				}
-
-				//claw
-				if(joystickGetDigital(1, 5, JOY_UP)) {
-		      motorSet(8, 127);
-		    }else if(joystickGetDigital(1, 5, JOY_DOWN)) {
-		      motorSet(8 ,-127);
-		    }else{
-		      motorSet(8, 0);
-		    }
+				//nick's controls
+				nDrive(); //nick's Drive
+				nLift();
+				nClaw();
 			}
 
-			//tip-bar code
-			if(joystickGetDigital(1, 7, JOY_UP)) {
-				tipSet(127); // pressing up, so tip-bar should open
-			}else if(joystickGetDigital(1, 7, JOY_DOWN)) {
-				tipSet(-127); // pressing down, so tip-bar should close
-			}else {
-				tipSet(0); // no buttons are pressed, stop the tip-bar
-			}
 
-			//control switch
-			if(joystickGetDigital(1, 8, JOY_UP)){
-				controller = 1;
-				delay(500);
-			}
+			//mutual controls
+			//these funtctions are located in "mutual.c"
+			tip(); // tip bar control
+			controlSwitch(controller);// control switcher
+			autoButton(); // autonomous activator button
+
 
       delay(20);
     }
